For lenet cnn use max-pooling. learning-rate independent of gpus.
Currently there is a mismatch in the n/w used for "Training on multiple GPUs from scratch" and for "Training on multiple GPUs with gluon". We should try to use the same pooling everywhere. Since lenet uses max-pooling using that as the standard here.

Second issue is that learning rate is made a multiple of gpus. Under certain cases that makes sense but under this case, it reduces the accuracy significantly which I don't think is a good idea for tutorials for beginners as it throws them off.