Minor word change to avoid triggering warnings.
MXNet integration tests of the Straight Dope (contained here:
http://jenkins.mxnet-ci.amazon-ml.com/job/NightlyTests_onBinaries/)
execute the notebook and line-scan for warnings and errors in the
output.

Kindly requesting we change 'Warning:' to 'Caution:' to avoid triggering
the log scan regular expression that normally picks up 'Warning:',
or else we'll require parsing the JSON in a more heavyweight manner that
seems unjustified at this point as there are no other triggers in this
book or the tutorials which also use the test framework.